refactor(tms): split the dashboard panel catalogue out of the settings service
dashboard-settings.service.ts was 413 lines, 200 of which were the
DEFAULT_PANELS literal. The catalogue moves to dashboard-panels.ts and the
service keeps behavior only, at 174 lines.

- hiddenByDefault and hiddenByDefaultIn were the same concept at two
  granularities; only hiddenByDefaultIn survives, with ALL_MODES covering
  "hidden everywhere". isHidden drops from a four-rung ?? chain to two.
- gatedPanels walks the list once and returns { visible, available }, so
  the two public computeds stop re-filtering a tuple array each.
- bottomOf() replaces the open-coded "row below the lowest panel" reduce
  that showPanel and mergeWithDefaults each carried.
- patchPanel() replaces three near-identical update + persist bodies.
- getPanel, updateAllPanels, the panels readonly and clearPersistedSettings
  had no callers anywhere and are gone. Nothing clears the persisted layout
  on logout now, but nothing did before either - the method was never wired
  up, and sidebar favorites and the onboarding-dismissed flag have the same
  gap. That wants one clear-local-state hook, not a method per service
  waiting for a caller.

One behavior fix rides along: mergeWithDefaults computed its stacking
offset across every kept panel, including the default-hidden ones sitting
at y=12..22. Gridster runs without compactType, so a panel introduced in a
later version would have landed below a dead band. Hidden panels no longer
count toward the offset, matching what showPanel already did.
